parrelize binary aggregation of functions
