+++ polars/polars-lazy/src/physical_plan/expressions/binary.rs
@@ -65,52 +65,6 @@ impl PhysicalExpr for BinaryExpr {
     }
 }
 
-impl PhysicalAggregation for BinaryFunctionExpr {
-    fn aggregate(
-        &self,
-        df: &DataFrame,
-        groups: &GroupTuples,
-        state: &ExecutionState,
-    ) -> Result<Option<Series>> {
-        let a = self.input_a.evaluate(df, state)?;
-        let b = self.input_b.evaluate(df, state)?;
-
-        let agg_a = a.agg_list(groups).expect("no data?");
-        let agg_b = b.agg_list(groups).expect("no data?");
-
-        // keep track of the output lengths. If they are all unit length,
-        // we can explode the array as it would have the same length as the no. of groups
-        // if it is not all unit length it should remain a listarray
-
-        let mut all_unit_length = true;
-
-        let ca = agg_a
-            .list()
-            .unwrap()
-            .into_iter()
-            .zip(agg_b.list().unwrap())
-            .map(|(opt_a, opt_b)| match (opt_a, opt_b) {
-                (Some(a), Some(b)) => {
-                    let out = self.function.call_udf(a, b).ok();
-
-                    if let Some(s) = &out {
-                        if s.len() != 1 {
-                            all_unit_length = false;
-                        }
-                    }
-                    out
-                }
-                _ => None,
-            })
-            .collect::<ListChunked>();
-
-        if all_unit_length {
-            return Ok(Some(ca.explode()?));
-        }
-        Ok(Some(ca.into_series()))
-    }
-}
-
 impl PhysicalAggregation for BinaryExpr {
     fn aggregate(
         &self,
